pyproject.toml
[project]
name = "barlow-track-simple"
version = "0.1.0"
description = "This repository contains functions that have been extracted or refactored to isolate the dependencies used in the original barlow-track"
readme = "README.md"
authors = [{ name = "Chung-Kuan Chen", email = "b97b01045@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "h5py",
    "hdbscan",
    "hdf5plugin",
    "matplotlib",
    "numpy",
    "polars",
    "ruamel-yaml",
    "tifffile",
    "torch",
    "torchio",
    "tqdm",
    "umap-learn",
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["ipykernel", "matplotlib", "ruff"]
